Extract uncolorme command into own module
